refactor: rearrange project tree
