+++ README.md
@@ -1,2 +1,4 @@
 # gwMonitor
-Monitor for PGW &amp; SGW
+Monitor for PGW & SGW
+
+Clone and move this project under $GOPATH/src/github.com/LinkerNetworks/ to start your work.
