+++ autoscaling/template.go
